+++ lib/cartero.rb
@@ -7,7 +7,7 @@ module Cartero
   PAYLOADS = {}
 
   def self.version
-    [0,5,2,"redirecto"]
+    [0,6,0,"mongolia"]
   end
 end
 
@@ -15,6 +15,7 @@ $LOAD_PATH.unshift(File.expand_path(File.dirname(__FILE__)))
 
 require 'cartero/base'
 require 'cartero/cli'
+#require 'cartero/repl'
 require 'cartero/db'
 require 'cartero/command'
 require 'cartero/cartero_patches'