+++ lib/cartero/models.rb
@@ -1,9 +1,9 @@
 #encoding: utf-8
 begin
-  require 'mongo_mapper'
+  require 'mongoid'
 rescue
   require 'active_support'
-  require 'mongo_mapper'
+  require 'mongoid'
 end
 
 $LOAD_PATH.unshift(File.dirname(__FILE__) + '/models')